# swagger/output/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "サポーターズ勉強会で使用するサンプルのAPIです\n"
  version: "1.0.0"
  title: "Supporterz API"
basePath: "/api"
schemes:
- "http"
produces:
- "application/json"
paths:
  /v1/users:
    get:
      tags:
      - "Supporterz"
      description: "ユーザーの一覧"
      operationId: "v1UsersGET"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SwUserListResponse"
      x-swagger-router-controller: "Supporterz"
    post:
      tags:
      - "Supporterz"
      description: "ユーザーの追加"
      operationId: "v1UsersPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/SwAddUser"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SwPostResponse"
      x-swagger-router-controller: "Supporterz"
  /v1/users/{id}:
    get:
      tags:
      - "Supporterz"
      description: "ユーザーの詳細"
      operationId: "v1UsersIdGET"
      parameters:
      - name: "id"
        in: "path"
        description: "ユーザーID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SwUserResponse"
      x-swagger-router-controller: "Supporterz"
    put:
      tags:
      - "Supporterz"
      description: "ユーザーの編集"
      operationId: "v1UsersIdPUT"
      parameters:
      - name: "id"
        in: "path"
        description: "ユーザーID"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/SwAddUser"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SwPostResponse"
      x-swagger-router-controller: "Supporterz"
    delete:
      tags:
      - "Supporterz"
      description: "ユーザーの削除"
      operationId: "v1UsersIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        description: "ユーザーID"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/SwPostResponse"
      x-swagger-router-controller: "Supporterz"
definitions:
  SwUserListResponse:
    type: "object"
    required:
    - "users"
    properties:
      users:
        type: "array"
        items:
          $ref: "#/definitions/SwUserListRow"
    example:
      users:
      - name: "name"
        id: 0
        email: "email"
      - name: "name"
        id: 0
        email: "email"
  SwUserListRow:
    required:
    - "email"
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
    example:
      name: "name"
      id: 0
      email: "email"
  SwUserResponse:
    type: "object"
    required:
    - "email"
    - "id"
    - "name"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
    example:
      name: "name"
      id: 0
      email: "email"
  SwAddUser:
    type: "object"
    required:
    - "email"
    - "name"
    properties:
      name:
        type: "string"
      email:
        type: "string"
    example:
      name: "name"
      email: "email"
  SwPostResponse:
    type: "object"
    required:
    - "status"
    properties:
      status:
        type: "string"
    example:
      status: "status"